UIHelpers: cast values to CGFloat

+++ UIKitExtensions/UIHelpers.h
@@ -9,9 +9,9 @@
 
 #define UI_NULL_VALUE_TO_BE_CHANGED_LATER 0
 
-#define UI_DESIGNATED_VALUE(value) value
-#define UI_CONTEXT_VALUE(value) value
-#define UI_CONTEXT_OFFSET(offset) offset
+#define UI_DESIGNATED_VALUE(value) (CGFloat)value
+#define UI_CONTEXT_VALUE(value) (CGFloat)value
+#define UI_CONTEXT_OFFSET(offset) (CGFloat)offset
 
 static inline UIView * topMostView() {
     return UIApplication.sharedApplication.windows.lastObject;
